requirements.txt
google-generativeai
pyttsx3
SpeechRecognition